+++ node/silkworm/recsplit/encoding/sequence.hpp
